fix circularity dependencies error
