+++ optrees/__init__.py
@@ -1,8 +1,7 @@
 # flake8: noqa
 
-from .graph.edge.edge import Edge
+from .graph.basic_objects import Edge, Vertex
 from .graph.graph import Graph
-from .graph.vertex.vertex import Vertex
 from .optimal_trees_algorithms.kruskal import kruskal
 from .optimal_trees_algorithms.prim import prim
 from .version import __version__
