+++ optrees/graph/basic_objects.py
@@ -1,7 +1,5 @@
 from typing import Union
 
-from optrees import Vertex
-
 
 class BasicEdge:
     def __init__(
@@ -67,8 +65,8 @@ class BasicEdge:
 
     @staticmethod
     def __get_start_vertex(
-        orientation: str, left_vertex: Vertex, right_vertex: Vertex
-    ) -> Union[None, Vertex]:
+        orientation: str, left_vertex: 'Vertex', right_vertex: 'Vertex'
+    ) -> Union[None, 'Vertex']:
         return (
             None
             if orientation not in ['->', '<-']
@@ -77,8 +75,8 @@ class BasicEdge:
 
     @staticmethod
     def __get_end_vertex(
-        orientation: str, left_vertex: Vertex, right_vertex: Vertex
-    ) -> Union[None, Vertex]:
+        orientation: str, left_vertex: 'Vertex', right_vertex: 'Vertex'
+    ) -> Union[None, 'Vertex']:
         return (
             None
             if orientation not in ['->', '<-']
@@ -90,11 +88,11 @@ class BasicEdge:
         return self.__label
 
     @property
-    def left_vertex(self) -> Vertex:
+    def left_vertex(self) -> 'Vertex':
         return self.__left_vertex
 
     @property
-    def right_vertex(self) -> Vertex:
+    def right_vertex(self) -> 'Vertex':
         return self.__right_vertex
 
     @property
@@ -114,11 +112,11 @@ class BasicEdge:
         self.__orientation = self.__validate_and_get_orientation(orientation)
 
     @property
-    def start(self) -> Vertex:
+    def start(self) -> 'Vertex':
         return self.__start
 
     @property
-    def end(self) -> Vertex:
+    def end(self) -> 'Vertex':
         return self.__end
 
 
@@ -127,3 +125,76 @@ class Edge(BasicEdge):
         self, left_vertex, right_vertex, weight=0.0, orientation='-', label=None
     ):
         super().__init__(left_vertex, right_vertex, weight, orientation, label)
+
+
+class BasicVertex:
+    def __init__(self, label: str):
+        self.__label = label
+        self.__edges: dict = {}
+        self.__neighbors: dict = {}
+
+    def __del__(self):
+        print(f'Vertex {self.label} is deleted.')
+
+    def __str__(self):
+        return self.__label
+
+    def __repr__(self):
+        return f'{self.__class__.__name__}({self.__label})'
+
+    def __eq__(self, other):
+        if isinstance(other, BasicVertex):
+            label_equal = self.__label == other.__label
+            edges_equal = self.__edges == other.__edges
+            neighbors_equal = self.__neighbors == other.__neighbors
+            return label_equal and edges_equal and neighbors_equal
+        return False
+
+    @property
+    def label(self) -> str:
+        return self.__label
+
+    @property
+    def edges(self) -> dict:
+        return self.__edges
+
+    @property
+    def neighbors(self) -> dict:
+        return self.__neighbors
+
+    def add_neighbor(self, vertex, weight=None, orientation='-'):
+        if not isinstance(vertex, BasicVertex):
+            raise TypeError('The vertex is not valid.')
+        same_vertex = self is vertex
+        already_neighbor = (
+            vertex.label in self.__neighbors.keys()
+            and self.__neighbors[vertex.label] is vertex
+        )
+        if not same_vertex and not already_neighbor:
+            self.__neighbors[vertex.label] = vertex
+            vertex.__neighbors[self.label] = self
+            edge = Edge(
+                label=f'{self.label} {orientation} {vertex.label}',
+                left_vertex=self,
+                right_vertex=vertex,
+                weight=weight,
+                orientation=orientation,
+            )
+            self.__edges[edge.label] = edge
+            vertex.__edges[edge.label] = edge
+        else:
+            raise ValueError('The vertex is already a neighbor.')
+
+
+class Vertex(BasicVertex):
+    def __init__(self, label: str):
+        super().__init__(label)
+
+    def amount_of_neighbors(self) -> int:
+        return len(super().neighbors)
+
+    def amount_of_edges(self) -> int:
+        return len(super().edges)
+
+    def is_isolated(self) -> bool:
+        return self.amount_of_neighbors() == 0
